--- src/main/java/ByeCommand.java
@@ -1,8 +1,0 @@
-public class ByeCommand extends Command {
-
-    @Override
-    public void execute(TaskList taskList, Terminal terminal) {
-        terminal.output("Buybye, see ya later...");
-        this.exit();
-    }
-}
