+++ src/main/java/demacia/commands/ByeCommand.java
@@ -1,0 +1,13 @@
+package demacia.commands;
+
+import demacia.tasks.TaskList;
+import demacia.ui.Terminal;
+
+public class ByeCommand extends Command {
+
+    @Override
+    public void execute(TaskList taskList, Terminal terminal) {
+        terminal.output("Buybye, see ya later...");
+        this.exit();
+    }
+}
