+++ src/main/java/demacia/commands/DeleteCommand.java
@@ -1,0 +1,27 @@
+package demacia.commands;
+
+import demacia.storage.SaveData;
+import demacia.tasks.TaskList;
+import demacia.ui.Terminal;
+
+public class DeleteCommand extends Command {
+
+    private final int index;
+
+    public DeleteCommand(int index) {
+        this.index = index;
+
+    }
+    @Override
+    public void execute(TaskList taskList, Terminal terminal) {
+        try {
+            String taskString = taskList.getTaskString(this.index);
+            taskList.deleteTask(this.index);
+            terminal.output("I have removed the task");
+            terminal.output(taskString);
+        } catch (IndexOutOfBoundsException e) {
+            terminal.output(e.getMessage());
+        }
+        this.save(new SaveData(taskList));
+    }
+}
