+++ src/main/java/demacia/exceptions/IncorrectArgumentFormatException.java
@@ -1,0 +1,12 @@
+package demacia.exceptions;
+
+public class IncorrectArgumentFormatException extends DukeException {
+
+    public IncorrectArgumentFormatException() {
+        super("Incorrect Argument Format");
+    }
+
+    public IncorrectArgumentFormatException(String msg) {
+        super("Incorrect Argument Format: \n" + msg);
+    }
+}
