+++ src/main/java/demacia/tasks/Deadline.java
@@ -1,3 +1,7 @@
+package demacia.tasks;
+
+import demacia.utils.Utils;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Date;
