+++ src/main/java/demacia/tasks/Event.java
@@ -1,3 +1,7 @@
+package demacia.tasks;
+
+import demacia.utils.Utils;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
