+++ src/main/java/demacia/utils/Utils.java
@@ -1,3 +1,4 @@
+package demacia.utils;
 import java.lang.StringBuilder;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
