--- src/main/java/exceptions/IncorrectArgumentFormatException.java
@@ -1,12 +1,0 @@
-package exceptions;
-
-public class IncorrectArgumentFormatException extends DukeException {
-
-    public IncorrectArgumentFormatException() {
-        super("Incorrect Argument Format");
-    }
-
-    public IncorrectArgumentFormatException(String msg) {
-        super("Incorrect Argument Format: \n" + msg);
-    }
-}
