+++ text-ui-test/runtest.bat
@@ -13,7 +13,7 @@ REM remove directory as well
 if exist data rmdir data
 
 REM compile the code into the out folder
-javac  -cp ..\src\main\java -Xlint:none -d ..\out ..\src\main\java\*.java
+javac  -cp ..\src\main\java -Xlint:none -d ..\out ..\src\main\java\demacia\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
@@ -21,7 +21,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\out Demacia < input.txt > ACTUAL.txt
+java -classpath ..\out demacia.Demacia < input.txt > ACTUAL.txt
 
 REM compare the output to the expected output
 FC ACTUAL.txt EXPECTED.txt
